Fix "public" conversion warnings
Ad explicit type conversions in some STL helpers. Add suffixes so that
we aren't egregiously mixing floats and doubles. This fixes
`-Wconversion` warnings in headers, so that users that use this warning
flag (and aren't treating our headers as system headers) aren't spammed
with warnings that really shouldn't exist.

There are a bunch of `-Wconversion` warnings in our sources as well that
really ought to also be fixed, but that can happen some other time.

+++ core/qtStlUtil.h
@@ -1,5 +1,5 @@
 /*ckwg +5
- * Copyright 2013 by Kitware, Inc. All Rights Reserved. Please refer to
+ * Copyright 2015 by Kitware, Inc. All Rights Reserved. Please refer to
  * KITWARE_LICENSE.TXT for licensing information, or contact General Counsel,
  * Kitware, Inc., 28 Corporate Drive, Clifton Park, NY 12065.
  */
@@ -38,7 +38,7 @@ inline QUrl qtUrl(const std::string& ss)
 inline std::string stdString(const QString& qs)
 {
   const QByteArray data = qs.toLocal8Bit();
-  return std::string(data.constData(), data.size());
+  return std::string(data.constData(), static_cast<size_t>(data.size()));
 }
 
 //-----------------------------------------------------------------------------
@@ -48,7 +48,7 @@ extern QTE_EXPORT std::string stdString(const QStringRef& qs);
 inline std::string stdString(const QStringRef& qs)
 {
   const QByteArray data = qs.toLocal8Bit();
-  return std::string(data.constData(), data.size());
+  return std::string(data.constData(), static_cast<size_t>(data.size()));
 }
 #endif
 
@@ -56,7 +56,7 @@ inline std::string stdString(const QStringRef& qs)
 inline std::string stdString(const QUrl& qu)
 {
   const QByteArray data = qu.toEncoded();
-  return std::string(data.constData(), data.size());
+  return std::string(data.constData(), static_cast<size_t>(data.size()));
 }
 
 //-----------------------------------------------------------------------------
